Extract retry logic to dedicated retry.go file
